Basic exercises for foundational Python learning

+++ app.py
@@ -1,4 +1,4 @@
-weight = input('How much do you weigh in (L)bs : ')
-converted = int(weight / 2.2)
+weight = float(input('How much do you weigh in (L)bs : '))
+converted = float(weight / 2.2)
 print(converted)
